Create readMe.txt
This subdir contains IDL codes that were used to perform some kind of check on an algorithm, or testing some aspect of CIAO/MARX to better understand how it worked.  Not all of these codes will be relevant to future work, but might serve as a good point of reference in case the same questions arise as what motivated these tests to begin with.